Allow to change event notification policy at company level

+++ app/models/membership.rb
@@ -36,17 +36,20 @@ class Membership < ActiveRecord::Base
     def create_notifications
       if memberable_type == 'Campaign' && company_user.role.has_permission?(:read, Campaign)
         Notification.new_campaign(company_user, memberable)
-      elsif memberable_type == 'Event' && company_user.allowed_to_access_place?(memberable.place)
-        Notification.new_event(company_user, memberable)
+      elsif memberable_type == 'Event' &&
+        memberable.company.setting(:event_alerts_policy, Notification::EVENT_ALERT_POLICY_TEAM) == Notification::EVENT_ALERT_POLICY_TEAM &&
+        company_user.allowed_to_access_place?(memberable.place)
+          Notification.new_event(company_user, memberable)
       end
     end
 
     def delete_notifications
       if memberable_type == 'Campaign'
         company_user.notifications.where(path: Rails.application.routes.url_helpers.campaign_path(memberable)).destroy_all
-      elsif memberable_type == 'Event'
-        company_user.notifications.where(path: Rails.application.routes.url_helpers.event_path(memberable)).destroy_all
-        company_user.notifications.where("params->'task_id' in (?)", memberable.task_ids.map{|n| n.to_s}).destroy_all
+      elsif memberable_type == 'Event' &&
+        memberable.company.setting(:event_alerts_policy, Notification::EVENT_ALERT_POLICY_TEAM) == Notification::EVENT_ALERT_POLICY_TEAM
+          company_user.notifications.where(path: Rails.application.routes.url_helpers.event_path(memberable)).destroy_all
+          company_user.notifications.where("params->'task_id' in (?)", memberable.task_ids.map{|n| n.to_s}).destroy_all
       end
     end
